% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prevent_equations.R
\name{app}
\alias{app}
\title{Navigate to browser-based Shiny implementation of PREVENT equations}
\usage{
app(...)
}
\arguments{
\item{...}{Not used. Reserved for future use.}
}
\value{
Returns \code{NULL} invisibly after opening app in your default browser.
}
\description{
This function opens a browser window (using the user's default browser) and
navigates to the Shiny app located at:

\url{https://martingmayer.shinyapps.io/prevent-equations}

Easier-to-remember URLs:
\itemize{
\item \url{https://tiny.cc/prevent-equations}
\item \url{https://tiny.cc/preventequations}
}

The app includes risk visualization and several options for customizing the output.
}
\examples{
app()
 
}
